Separate messages subdirs into files
