+++ lib/http/commands/controls.rb
@@ -5,6 +5,9 @@ require 'connection/controls'
 
 require 'http/commands/controls/dialogs'
 require 'http/commands/controls/messages'
+require 'http/commands/controls/messages/requests'
+require 'http/commands/controls/messages/resources'
+require 'http/commands/controls/messages/responses'
 require 'http/commands/controls/run_server'
 require 'http/commands/controls/server'
 require 'http/commands/controls/uri'
